+++ api/src/controllers/user.js
@@ -1,4 +1,4 @@
-const { User } = require("../db.js")
+const { User, Review } = require("../db.js")
 const bcrypt = require("bcryptjs")
 const { generateToken } = require("../helpers/token.js")
 const { Op } = require("sequelize")
@@ -7,6 +7,7 @@ const { verifyToken } = require("../helpers/verify.js")
 const { simpleToken } = require("../helpers/simpleToken.js")
 const { generatePassword } = require("../helpers/generatePassword.js")
 const { emailForgotPassword } = require("../helpers/emailForgotPassword.js")
+const { getTestMessageUrl } = require("nodemailer")
 
 module.exports = {
   userSingUp: async (req, res) => {
@@ -53,12 +54,12 @@ module.exports = {
   },
 
   userSingUpOrSingInGoogle: async (req, res) => {
-    const {email, username} = req.body;
+    const { email, username } = req.body
     try {
       const user = await User.findOne({
-        where: {email}
-      });
-      if(user){
+        where: { email },
+      })
+      if (user) {
         const token = generateToken({ id: user.id, isAdmin: user.isAdmin })
         return res.status(200).send({ token, admin: user.isAdmin })
       }
@@ -70,19 +71,17 @@ module.exports = {
       })
       const token = generateToken({ id: newUser.id, isAdmin: newUser.isAdmin })
       return res.status(200).send({ token, admin: newUser.isAdmin })
-    } catch (error) {
-      
-    }
+    } catch (error) {}
   },
 
   forgotPassword: async (req, res) => {
     const { email } = req.body
+
     try {
       const userExists = await User.findOne({ where: { email } })
       if (!userExists) {
-        throw new Error("User does not exist")
+        return res.send({ status: false })
       }
-
       userExists.token = simpleToken() // Agregar token al modelo User? y una funcion que genere un Token
       await userExists.save()
 
@@ -105,11 +104,7 @@ module.exports = {
       const user = await User.findOne({ where: { token } })
 
       if (!user) {
-        throw new Error("User does not exist")
-      }
-
-      if (password.length < 4) {
-        throw new Error("Password must have more than 4 characters")
+        return res.status(201).send({ status: false })
       }
 
       user.token = null
@@ -142,7 +137,7 @@ module.exports = {
         throw new Error("Password must have more than 4 characters")
       }
       const saltRounds = 10
-      const passwordHash = await bcrypt.hash(password, saltRounds)
+      const passwordHash = await bcrypt.hash(newPassword, saltRounds)
 
       user.password = passwordHash
       user.save()
@@ -161,6 +156,18 @@ module.exports = {
       sendError(res, error)
     }
   },
+  getSuperAdmin: async (req, res) => {
+    try {
+      const decodedToken = await verifyToken(req, res)
+      const admin = await User.findByPk(decodedToken.id)
+      if (admin.email === "admin@gmail.com") {
+        return res.send({ superAdmin: true })
+      }
+      return res.send({ superAdmin: false })
+    } catch (error) {
+      sendError(res, error)
+    }
+  },
   getAllUsers: async (req, res) => {
     try {
       const { search = "", isAdmin } = req.query
@@ -175,7 +182,7 @@ module.exports = {
           },
         })
 
-        return res.status(302).send(usuarioSearched)
+        return res.status(200).send(usuarioSearched)
       }
       if (search.length) {
         const usuarioSearched = await User.findAll({
@@ -187,7 +194,7 @@ module.exports = {
           },
         })
 
-        return res.status(302).send(usuarioSearched)
+        return res.status(200).send(usuarioSearched)
       }
       const allUsers = await User.findAll()
       return res.status(200).send(allUsers)
@@ -197,13 +204,21 @@ module.exports = {
   },
   deleteUser: async (req, res) => {
     try {
-      const { email } = req.params
-      const removedUser = await User.destroy({
-        where: { email },
+      const decodedToken = await verifyToken(req, res)
+      const userWillingToDelete = await User.findOne({
+        where: { id: decodedToken.id },
       })
-      if (removedUser) return res.send({ msg: `User ${email} removed` })
-      return res.status(400).send({
-        error: `User ${email} doesnt exist`,
+      const { email } = req.params
+      if (userWillingToDelete.dataValues.email === "admin@gmail.com") {
+        const user = await User.findOne({ where: { email } })
+        await Review.destroy({ where: { userId: user.id } })
+        await User.destroy({
+          where: { email },
+        })
+        return res.send({ msg: `User ${email} removed` })
+      }
+      return res.send({
+        msg: `User ${userWillingToDelete.dataValues.email} has no permitions to delete other users`,
       })
     } catch (error) {
       sendError(res, error)
@@ -211,19 +226,28 @@ module.exports = {
   },
   changeUsersRole: async (req, res) => {
     try {
-      const { email, adminState } = req.body
-      const foundUser = await User.findOne({
-        where: { email },
+      const decodedToken = await verifyToken(req, res)
+      const userWillingToChangeRole = await User.findOne({
+        where: { id: decodedToken.id },
       })
-      if (foundUser) {
-        foundUser.isAdmin = adminState
-        foundUser.save()
-        return res.send(
-          `${email} was changed to ${adminState ? "Admin" : "User"}`
-        )
-      } else {
-        return res.send("The email passed was not found")
+      const { email, adminState } = req.body
+      if (userWillingToChangeRole.dataValues.email === "admin@gmail.com") {
+        const foundUser = await User.findOne({
+          where: { email },
+        })
+        if (foundUser) {
+          foundUser.isAdmin = adminState
+          foundUser.save()
+          return res.send(
+            `${email} was changed to ${adminState ? "Admin" : "User"}`
+          )
+        } else {
+          return res.send("The email passed was not found")
+        }
       }
+      return res.send({
+        msg: `User ${userWillingToChangeRole.dataValues.email} has no permitions to change other users role`,
+      })
     } catch (error) {
       return sendError(res, error)
     }
@@ -232,7 +256,9 @@ module.exports = {
     try {
       const decodedToken = await verifyToken(req, res)
       const id = decodedToken.id
-      const userName = await User.findByPk(id, { attributes: ["userName","email"] })
+      const userName = await User.findByPk(id, {
+        attributes: ["userName", "email"],
+      })
       res.send(userName)
     } catch (error) {
       sendError(res, error)
